// src/gen/tvm-spec.d.ts
export type InstructionName = string;
export type CategoryOfInstruction = string;
/**
 * Free-form markdown description of instruction.
 */
export type InstructionDescription = string;
/**
 * Free-form description of gas amount used by instruction.
 */
export type GasUsageInfo = string;
/**
 * Free-form fift usage description.
 */
export type FiftUsageDoc = string;
export type FiftSnippet = string;
export type ExampleDescription = string;
/**
 * Free-form bytecode format description.
 */
export type OpcodeFormatDocumentation = string;
/**
 * TL-b bytecode format description.
 */
export type TLBSchema = string;
/**
 * Prefix to determine next instruction to parse. It is a hex bitstring as in TL-b (suffixed with `_` if bit length is not divisible by 4, trailing `'1' + '0' * x` must be removed).
 */
export type InstructionPrefix = string;
// ...
export type InstructionOperands = {
  name: OperandVariableName;
  loader: LoaderFunctionForOperand;
  loader_args?: ArgumentsForLoaderFunctionOptionalNoArgumentsInCaseOfAbsence;
}[];
/**
 * Free-form description of stack inputs and outputs. Usually the form is `[inputs] - [outputs]` where `[inputs]` are consumed stack values and `outputs` are produced stack values (top of stack is the last value).
 */
export type StackUsageDescription = string;
// ...
export interface Instruction {
  mnemonic: InstructionName;
  doc: Documentation;
  bytecode: BytecodeFormat;
  value_flow?: ValueFlowOfInstruction;
}
/**
 * Free-form human-friendly information which should be used for documentation purposes only.
 */
export interface Documentation {
  category: CategoryOfInstruction;
  description?: InstructionDescription;
  gas?: GasUsageInfo;
  fift?: FiftUsageDoc;
  fift_examples?: {
    fift?: FiftSnippet;
    description?: ExampleDescription;
  }[];
}
/**
 * Information related to bytecode format of an instruction. Assuming that each instruction has format `prefix || operand_1 || operand_2 || ...` (also some operands may be refs, not bitstring part).
 */
export interface BytecodeFormat {
  doc_opcode?: OpcodeFormatDocumentation;
  tlb: TLBSchema;
  prefix: InstructionPrefix;
  operands_range_check?: OperandsRangeCheck;
  operands?: InstructionOperands;
}
/**
 * In TVM, it is possible for instructions to have overlapping prefixes, so to determine actual instruction it is required to read next `length` bits after prefix as uint `i` and check `from <= i <= to`. Optional, there is no operands check in case of absence.
 */
export interface OperandsRangeCheck {
  length: number;
  from: number;
  to: number;
}
export interface ArgumentsForLoaderFunctionOptionalNoArgumentsInCaseOfAbsence {
  [k: string]: unknown;
}
/**
 * Information related to usage of stack and registers by instruction.
 */
export interface ValueFlowOfInstruction {
  doc_stack?: StackUsageDescription;
  inputs?: InstructionInputs;
  outputs?: InstructionOutputs;
}
/**
 * Incoming values constraints. Input is unconstrained if absent.
 */
export interface InstructionInputs {
  stack: StackValues;
}
/**
 * Outgoing values constraints. Output is unconstrained if absent.
 */
export interface InstructionOutputs {
  stack: StackValues;
}